pyproject.toml
[project]
name = "pipecat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.18",
    "bv>=0.1.4",
    "colorama>=0.4.6",
    "deepgram-sdk>=4.1.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "google-cloud>=0.34.0",
    "google-cloud-speech>=2.32.0",
    "google-genai>=1.14.0",
    "groq>=0.24.0",
    "librosa>=0.11.0",
    "lightning-whisper-mlx>=0.0.10",
    "mlx>=0.25.1",
    "opencv-python>=4.11.0.86",
    "pipecat-ai[noisereduce,silero]>=0.0.67",
    "simli-ai>=0.1.11",
    "sounddevice>=0.5.1",
    "torchaudio>=2.7.0",
    "uv>=0.7.3",
    "uvicorn>=0.34.2",
    "xmltodict>=0.14.2",
    "zmq>=0.0.0",
]
